+++ src/test/java/lotto/domain/LottoManagerTest.java
@@ -1,5 +1,6 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.LottoManager;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
