+++ src/test/java/lotto/domain/LottoNumberTest.java
@@ -1,11 +1,10 @@
-package lotto;
+package lotto.domain;
 
 import lotto.exception.InvalidBoundLottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 class LottoNumberTest {
